Edit Workout Functionality Test
