--- client/src/components/WorkoutsView/WorkoutTitleList.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-const WorkoutTitleList = props => {
-  return <div>{props.exerciseObj.name}</div>;
-};
-
-export default WorkoutTitleList;
