+++ client/src/components/WorkoutsView/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import requireAuth from '../../requireAuth';
-import AddWorkouts from './Semantic/AddWorkouts';
-import MyWorkouts from './Semantic/MyWorkouts';
-import EditWorkout from './Semantic/EditWorkout';
+import AddWorkouts from './AddWorkouts';
+import MyWorkouts from './MyWorkouts';
+import EditWorkout from './EditWorkout';
 import styled from 'styled-components';
 
 const WorkoutsView = props => {
